Extract Protocol Kit Safe types
