+++ packages/protocol-kit/src/Safe.ts
@@ -1,7 +1,6 @@
 import { BigNumber } from '@ethersproject/bignumber'
 import {
   EthAdapter,
-  MetaTransactionData,
   OperationType,
   SafeMultisigTransactionResponse,
   SafeTransaction,
@@ -11,19 +10,26 @@ import {
   TransactionOptions,
   TransactionResult
 } from '@safe-global/safe-core-sdk-types'
-import {
-  SAFE_FEATURES,
-  hasSafeFeature,
-  isMetaTransactionArray,
-  isSafeMultisigTransactionResponse,
-  sameString
-} from './utils'
 import ContractManager from './managers/contractManager'
 import FallbackHandlerManager from './managers/fallbackHandlerManager'
 import GuardManager from './managers/guardManager'
 import ModuleManager from './managers/moduleManager'
 import OwnerManager from './managers/ownerManager'
-import { ContractNetworksConfig } from './types'
+import {
+  AddOwnerTxParams,
+  ConnectSafeConfig,
+  CreateTransactionProps,
+  RemoveOwnerTxParams,
+  SafeConfig,
+  SwapOwnerTxParams
+} from './types'
+import {
+  hasSafeFeature,
+  isMetaTransactionArray,
+  isSafeMultisigTransactionResponse,
+  SAFE_FEATURES,
+  sameString
+} from './utils'
 import {
   generateEIP712Signature,
   generatePreValidatedSignature,
@@ -37,58 +43,6 @@ import {
   standardizeMetaTransactionData,
   standardizeSafeTransactionData
 } from './utils/transactions/utils'
-
-export interface SafeConfig {
-  /** ethAdapter - Ethereum adapter */
-  ethAdapter: EthAdapter
-  /** safeAddress - The address of the Safe account to use */
-  safeAddress: string
-  /** isL1SafeMasterCopy - Forces to use the GnosisSafe L1 version of the contract instead of the L2 version */
-  isL1SafeMasterCopy?: boolean
-  /** contractNetworks - Contract network configuration */
-  contractNetworks?: ContractNetworksConfig
-}
-
-export interface ConnectSafeConfig {
-  /** ethAdapter - Ethereum adapter */
-  ethAdapter?: EthAdapter
-  /** safeAddress - The address of the Safe account to use */
-  safeAddress?: string
-  /** isL1SafeMasterCopy - Forces to use the GnosisSafe L1 version of the contract instead of the L2 version */
-  isL1SafeMasterCopy?: boolean
-  /** contractNetworks - Contract network configuration */
-  contractNetworks?: ContractNetworksConfig
-}
-
-export interface CreateTransactionProps {
-  /** safeTransactionData - The transaction or transaction array to process */
-  safeTransactionData: SafeTransactionDataPartial | MetaTransactionData[]
-  /** options - The transaction array optional properties */
-  options?: SafeTransactionOptionalProps
-  /** onlyCalls - Forces the execution of the transaction array with MultiSendCallOnly contract */
-  onlyCalls?: boolean
-}
-
-export interface AddOwnerTxParams {
-  /** ownerAddress - The address of the new owner */
-  ownerAddress: string
-  /** threshold - The new threshold */
-  threshold?: number
-}
-
-export interface RemoveOwnerTxParams {
-  /** ownerAddress - The address of the owner that will be removed */
-  ownerAddress: string
-  /** threshold - The new threshold */
-  threshold?: number
-}
-
-export interface SwapOwnerTxParams {
-  /** oldOwnerAddress - The old owner address */
-  oldOwnerAddress: string
-  /** newOwnerAddress - The new owner address */
-  newOwnerAddress: string
-}
 
 class Safe {
   #ethAdapter!: EthAdapter
